Added validation to ads posted on dashboard

+++ resources/views/home.blade.php
@@ -52,34 +52,46 @@
 
                     </div>
                     <div class="tab-pane fade" id="nav-profile" role="tabpanel" aria-labelledby="nav-profile-tab">
-                        <table class="table">
-                            <thead>
-                            <tr>
-                                <th scope="col" style="width: 10%">Title</th>
-                                <th scope="col" style="width: 10%">Price</th>
-                                <th scope="col" style="width: 45%">Description</th>
-                                <th scope="col" style="width: 15%">Created At</th>
-                                <th scope="col" style="width: 20%">Actions</th>
-                            </tr>
-                            </thead>
-                            <tbody>
-                                @foreach($posts as $key => $post)
 
-                                        <tr>
-                                            <td>{{$post->title}}</td>
-                                            <td>${{ $post->price }}</td>
-                                            <td>{{ substr($post->description, 0, 50) }}</td>
-                                            <td>{{ date('M j, Y H:i', strtotime($post->created_at)) }}</td>
-                                            <td>
-                                                <a class="btn btn-secondary" href="/ad/{{ $post->id }}" role="button">View</a>
-                                                <a href="/ad/{{ $post->id }}/edit"><button type="button" class="btn btn-info">Edit</button></a>
-                                                <a href="/ad/{{ $post->id }}/destroy"><button type="button" class="btn btn-danger">Delete</button></a>
-                                            </td>
-                                        </tr>
+                        @if (count($posts)>0)
+                            <table class="table">
+                                <thead>
+                                <tr>
+                                    <th scope="col" style="width: 10%">Title</th>
+                                    <th scope="col" style="width: 10%">Price</th>
+                                    <th scope="col" style="width: 45%">Description</th>
+                                    <th scope="col" style="width: 15%">Created At</th>
+                                    <th scope="col" style="width: 20%">Actions</th>
+                                </tr>
+                                </thead>
+                                <tbody>
+                                        @foreach($posts as $key => $post)
 
-                                @endforeach
-                                </tbody>
+                                                <tr>
+                                                    <td>{{$post->title}}</td>
+                                                    <td>${{ $post->price }}</td>
+                                                    <td>{{ substr($post->description, 0, 50) }}</td>
+                                                    <td>{{ date('M j, Y H:i', strtotime($post->created_at)) }}</td>
+                                                    <td>
+                                                        <a class="btn btn-secondary" href="/ad/{{ $post->id }}" role="button">View</a>
+                                                        <a href="/ad/{{ $post->id }}/edit"><button type="button" class="btn btn-info">Edit</button></a>
+                                                        <a href="/ad/{{ $post->id }}/destroy"><button type="button" class="btn btn-danger">Delete</button></a>
+                                                    </td>
+                                                </tr>
+
+                                        @endforeach
+
+                                    </tbody>
                             </table>
+
+                        @else
+
+                            <div class="text-center">
+                                <h3>No ads posted.</h3>
+                                <a class="btn btn-primary btn-lg" href="/ad/create" role="button">Post New Ad</a>
+                            </div>
+
+                        @endif
                     </div>
                 </div>
